Split checkout phase out of executor.go
